# src/swagger.json
{
    "openapi": "3.0.0-rc2",
...
    "info": {
        "title": "Gstock API",
        "version": "3.0.0",
        "description": "Gstock API under development",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "davidtseng1992@hotmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [{
        "url": "{schemes}://{host}/{endpoints}",
        "description": "Admin server",
        "variables": {
            "host": {
                "enum": [
                    "localhost",
                    "gstock.sg"
                ],
                "default": "gstock.sg"
            },
            "schemes": {
                "enum": [
                    "https",
                    "http"
                ],
                "default": "https"
            },
            "endpoints": {
                "enum": [
                    "admin.gstock",
                    "buyer.gstock",
                    "merchant.gstock"
                ],
                "default": "https"
            }
        }
    }],
    "security": [{
        "jwt_auth": []
    }],
...
    "paths": {
...
        "/merchant/product": {
            "get": {
                "description": "Get products belong to a merchant",
                "operationId": "merchantGetProduct",
                "tags": [
                    "Merchant"
                ],
                "parameters": [{
                        "$ref": "#/components/parameters/pageParam"
                    },
                    {
                        "$ref": "#/components/parameters/limitParam"
                    },
                    {
                        "$ref": "#/components/parameters/productNameParam"
                    },
                    {
                        "$ref": "#/components/parameters/productPriceLowParam"
                    },
                    {
                        "$ref": "#/components/parameters/productPriceHighParam"
                    },
                    {
                        "$ref": "#/components/parameters/productCategoryParam"
                    },
                    {
                        "$ref": "#/components/parameters/productProductIdParam"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated!"
                    }
                },
                "security": [{
                    "jwt_auth": []
                }]
            },
            "post": {
                "description": "Post a new product with basic detail.",
                "operationId": "merchantCreateProduct",
                "tags": [
                    "Merchant"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StartProductCreation"
                            }
                        }
                    }
                },
                "security": [{
                    "jwt_auth": []
                }],
                "responses": {
                    "200": {
                        "description": "Product created!"
                    },
                    "409": {
                        "description": "Product name exists!"
                    }
                }
            }
        },
...
        "/merchant/product/{productId}": {
            "put": {
                "description": "Update a product detail by productId.",
                "operationId": "merchantUpdateProduct",
                "tags": [
                    "Merchant"
                ],
                "parameters": [{
                    "$ref": "#/components/parameters/productProductIdPath"
                }],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BaseProduct"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product created!"
                    },
                    "409": {
                        "description": "Product name exists!"
                    }
                }
            }
        },
        "/merchant/variant/{productId}": {
            "post": {
                "description": "Create a product variant given product id. You can submit a single object or an array of object.",
                "operationId": "merchantCreateVariant",
                "tags": [
                    "Merchant"
                ],
                "parameters": [{
                    "$ref": "#/components/parameters/productProductIdPath"
                }],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VariantDetail"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Variant added!"
                    },
                    "409": {
                        "description": "Variant exist!"
                    }
                }
            },
            "put": {
                "description": "Update a product variant given variantId",
                "tags": [
                    "Merchant"
                ],
                "parameters": [{
                    "$ref": "#/components/parameters/productProductIdPath"
                }],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VariantDetail"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Variant added!"
                    },
                    "409": {
                        "description": "Variant exist!"
                    }
                }
            }
        },
        "/merchant/variant/option/{productId}": {
            "get": {
                "description": "Get a product variant's all options and",
                "operationId": "merchantGetAllVariantOptions",
                "tags": [
                    "Merchant"
                ],
                "parameters": [{
                    "$ref": "#/components/parameters/productProductIdPath"
                }],
                "responses": {
                    "200": {
                        "description": "Options!"
                    }
                }
            }
        },
        "/merchant/variant/delete": {
            "post": {
                "description": "Delete an existing product variant given variant ID.",
                "operationId": "merchantDeleteVariant",
                "tags": [
                    "Merchant"
                ],
                "parameters": [],
                "security": [{
                    "jwt_auth": []
                }],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductVariantID"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Options!"
                    }
                }
            }
        },
        "/merchant/thumbnail/delete/{productId}": {
            "post": {
                "description": "Delete all product picture given a productId",
                "operationId": "merchantRemoveProductImage",
                "tags": [
                    "Merchant"
                ],
                "parameters": [{
                    "$ref": "#/components/parameters/productProductIdPath"
                }],
                "security": [{
                    "jwt_auth": []
                }],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductImagesName"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Variant!"
                    }
                }
            }
        },
        "/merchant/thumbnail/{productId}": {
            "get": {
                "description": "Get product images name.",
                "operationId": "merchantGetImageUrl",
                "tags": [
                    "Merchant"
                ],
                "parameters": [{
                    "$ref": "#/components/parameters/productProductIdPath"
                }],
                "responses": {
                    "200": {
                        "description": "Images"
                    }
                }
            }
        },
        "/product/product": {
            "get": {
                "description": "Get single product category.",
                "tags": [
                    "Product"
                ],
                "parameters": [{
                    "$ref": "#/components/parameters/productProductIdParam"
                }],
                "responses": {
                    "200": {
                        "description": "Variant added!"
                    },
                    "409": {
                        "description": "Variant exist!"
                    }
                }
            }
        },
...
        "/product/category": {
            "get": {
                "description": "Get all product category.",
                "tags": [
                    "Product"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Variant added!"
                    },
                    "409": {
                        "description": "Variant exist!"
                    }
                }
            }
        },
...
        "/merchant/variant_options": {
            "get": {
                "description": "Get an attribute's options given attribute title or the option given its option name.",
                "tags": [
                    "Attribute"
                ],
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Variant added!"
                    },
                    "409": {
                        "description": "Variant exist!"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt_auth": {
                "description": "JWT auth",
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "bearer"
            }
        },
        "parameters": {
            "productProductIdPath": {
                "in": "path",
                "name": "productId",
                "schema": {
                    "type": "string"
                },
                "required": true,
                "description": "Product Id"
            },
...
            "productNameParam": {
                "description": "Search product based on name.",
                "in": "query",
                "name": "name",
                "schema": {
                    "type": "string"
                }
            },
            "productPriceLowParam": {
                "description": "Search product price lower than.",
                "in": "query",
                "name": "low",
                "schema": {
                    "type": "integer"
                }
            },
            "productPriceHighParam": {
                "description": "Search product price higher than.",
                "in": "query",
                "name": "high",
                "schema": {
                    "type": "integer"
                }
            },
            "productCategoryParam": {
                "description": "Search product based on category.",
                "in": "query",
                "name": "category",
                "schema": {
                    "type": "string"
                }
            },
...
            "productProductIdParam": {
                "description": "Product id.",
                "in": "query",
                "name": "product_id",
                "schema": {
                    "type": "string"
                }
            },
...
            "pageParam": {
                "description": "The number of items to skip before starting to collect the result set.",
                "in": "query",
                "name": "page",
                "required": true,
                "schema": {
                    "type": "integer",
                    "minimum": 1
                },
                "example": 1
            },
            "limitParam": {
                "in": "query",
                "name": "limit",
                "required": true,
                "schema": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 50
                },
                "example": 20,
                "description": "The numbers of items to return"
            }
        },
        "schemas": {
...
            "StartProductCreation": {
                "description": "Starting product creation detail",
                "type": "object",
                "properties": {
                    "category": {
                        "type": "string"
                    }
                }
            },
            "BaseProduct": {
                "description": "Product basic detail",
                "type": "object",
                "properties": {
                    "detail": {
                        "$ref": "#/components/schemas/ProductDetail"
                    },
                    "brief": {
                        "$ref": "#/components/schemas/ProductBrief"
                    },
                    "pricing": {
                        "$ref": "#/components/schemas/ProductPricing"
                    },
                    "stock": {
                        "$ref": "#/components/schemas/ProductStock"
                    }
                }
            },
            "ProductImagesName": {
                "type": "object",
                "properties": {
                    "remove": {
                        "type": "string"
                    }
                }
            },
            "ProductDetail": {
                "type": "object",
                "properties": {
                    "long_description": {
                        "type": "string"
                    },
                    "product_brand": {
                        "type": "string"
                    },
                    "barcode": {
                        "type": "string"
                    },
                    "sku": {
                        "type": "string"
                    }
                }
            },
            "ProductPricing": {
                "type": "object",
                "properties": {
                    "final_price": {
                        "type": "number"
                    },
                    "min_price": {
                        "type": "number"
                    },
                    "max_price": {
                        "type": "number"
                    },
                    "discount_price": {
                        "type": "number"
                    }
                }
            },
            "ProductStock": {
                "type": "object",
                "properties": {
                    "qty": {
                        "type": "number"
                    },
                    "min_qty": {
                        "type": "number"
                    }
                }
            },
            "ProductBrief": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "short_description": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "discount_price": {
                        "type": "number"
                    },
                    "category": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    }
                }
            },
            "VariantDetail": {
                "description": "Variant details.",
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "flavor/size/color"
                    },
                    "value": {
                        "type": "string",
                        "example": "bitter/sweet/red"
                    },
                    "price": {
                        "type": "number",
                        "example": 4
                    },
                    "discount_price": {
                        "type": "number",
                        "example": 6
                    },
                    "image": {
                        "type": "string",
                        "example": "big.jpg"
                    },
                    "stock": {
                        "type": "number",
                        "example": 1230
                    },
                    "sku": {
                        "type": "string",
                        "example": "ASD - 1231"
                    }
                }
            },
...
            "ProductVariantID": {
                "description": "Cart details.",
                "type": "object",
                "properties": {
                    "variant_id": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}